deepfake-detection/backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
opencv-python-headless
facenet-pytorch
numpy
timm
pillow
pytest
httpx
